+++ extensions/chatgpt/src/views/empty.tsx
@@ -1,9 +1,13 @@
 import { Icon, List } from "@raycast/api";
+import { getConfiguration } from "../hooks/useChatGPT";
 
-export const EmptyView = () => (
-  <List.EmptyView
-    title="Ask anything!"
-    description={"Type your question or prompt from the search bar and hit the enter key"}
-    icon={Icon.QuestionMark}
-  />
-);
+export const EmptyView = () => {
+  const { provider } = getConfiguration();
+  return (
+    <List.EmptyView
+      title={`Ask ${provider} anything!`}
+      description={"Type your question or prompt from the search bar and hit the enter key"}
+      icon={Icon.QuestionMark}
+    />
+  );
+};