+++ extensions/chatgpt/src/views/model/list.tsx
@@ -52,6 +52,7 @@ const ModelDetailView = (props: { model: Model; markdown?: string | null | undef
             <List.Item.Detail.Metadata.TagList.Item text={model.option} />
           </List.Item.Detail.Metadata.TagList>
           <List.Item.Detail.Metadata.Label title="Temperature" text={model.temperature.toLocaleString()} />
+          {model.provider && <List.Item.Detail.Metadata.Label title="Provider" text={model.provider} />}
           <List.Item.Detail.Metadata.Separator />
           <List.Item.Detail.Metadata.Label title="ID" text={model.id} />
           <List.Item.Detail.Metadata.Label title="Updated at" text={new Date(model.updated_at ?? 0).toLocaleString()} />